refactor xcm_config.rs: * move a part of the code in the litmus/moonbase/rococo xcm_config.rs(struct: FirstAssetTrader XcmFeesToAccount MultiNativeAsset) to runtime_common module:
