+++ runtime/rococo/src/xcm_config.rs
@@ -16,6 +16,38 @@
 #![allow(clippy::clone_on_copy)]
 #![allow(clippy::useless_conversion)]
 
+use codec::{Decode, Encode};
+use frame_support::{
+	match_types, parameter_types,
+	traits::{Everything, Nothing, PalletInfoAccess},
+	weights::{IdentityFee, Weight},
+	PalletId,
+};
+use orml_traits::{location::AbsoluteReserveProvider, parameter_type_with_key};
+use pallet_xcm::XcmPassthrough;
+use polkadot_parachain::primitives::Sibling;
+use scale_info::TypeInfo;
+// Litentry: The CheckAccount implementation is forced by the bug of FungiblesAdapter.
+// We should replace () regarding fake_pallet_id account after our PR passed.
+use sp_runtime::traits::{AccountIdConversion, Convert as spConvert};
+use sp_std::{borrow::Borrow, prelude::*};
+use xcm::latest::prelude::*;
+use xcm_builder::{
+	AccountId32Aliases, AllowTopLevelPaidExecutionFrom, AllowUnpaidExecutionFrom,
+	ConvertedConcreteAssetId, CurrencyAdapter, EnsureXcmOrigin, FixedWeightBounds,
+	FungiblesAdapter, IsConcrete, LocationInverter, ParentIsPreset, RelayChainAsNative,
+	SiblingParachainAsNative, SiblingParachainConvertsVia, SignedAccountId32AsNative,
+	SignedToAccountId32, SovereignSignedViaLocation, TakeWeightCredit, UsingComponents,
+};
+use xcm_executor::{
+	traits::{Convert as xcmConvert, JustTry},
+	XcmExecutor,
+};
+
+use pallet_asset_manager::AssetTypeGetter;
+use primitives::AccountId;
+use runtime_common::xcm_impl::{FirstAssetTrader, MultiNativeAsset, XcmFeesToAccount};
+
 #[cfg(test)]
 use super::tests::setup::ParachainXcmRouter;
 use super::{
@@ -24,38 +56,6 @@ use super::{
 };
 #[cfg(not(test))]
 use super::{ParachainSystem, XcmpQueue};
-
-use frame_support::{
-	match_types, parameter_types,
-	traits::{tokens::fungibles::Mutate, Everything, Get, Nothing, PalletInfoAccess},
-	weights::{constants::WEIGHT_PER_SECOND, IdentityFee, Weight},
-	PalletId,
-};
-use orml_traits::{location::AbsoluteReserveProvider, parameter_type_with_key};
-use pallet_asset_manager::{AssetTypeGetter, UnitsToWeightRatio};
-use pallet_xcm::XcmPassthrough;
-use polkadot_parachain::primitives::Sibling;
-use primitives::AccountId;
-use sp_runtime::traits::Zero;
-use sp_std::marker::PhantomData;
-use xcm::latest::{prelude::*, AssetId as xcmAssetId};
-use xcm_builder::{
-	AccountId32Aliases, AllowTopLevelPaidExecutionFrom, AllowUnpaidExecutionFrom,
-	ConvertedConcreteAssetId, CurrencyAdapter, EnsureXcmOrigin, FixedWeightBounds,
-	FungiblesAdapter, IsConcrete, LocationInverter, ParentIsPreset, RelayChainAsNative,
-	SiblingParachainAsNative, SiblingParachainConvertsVia, SignedAccountId32AsNative,
-	SignedToAccountId32, SovereignSignedViaLocation, TakeRevenue, TakeWeightCredit,
-	UsingComponents,
-};
-use xcm_executor::{
-	traits::{Convert as xcmConvert, FilterAssetLocation, JustTry, MatchesFungibles, WeightTrader},
-	XcmExecutor,
-};
-
-use codec::{Decode, Encode};
-use scale_info::TypeInfo;
-use sp_runtime::traits::Convert as spConvert;
-use sp_std::{borrow::Borrow, prelude::*};
 
 parameter_types! {
 	pub const RelayLocation: MultiLocation = MultiLocation::parent();
@@ -112,9 +112,6 @@ pub type LocalAssetTransactor = CurrencyAdapter<
 	(),
 >;
 
-// Litentry: The CheckAccount implementation is forced by the bug of FungiblesAdapter.
-// We should replace () regarding fake_pallet_id account after our PR passed.
-use sp_runtime::traits::AccountIdConversion;
 parameter_types! {
 	pub const TempPalletId: PalletId = PalletId(*b"py/tempA");
 	pub TempAccount: AccountId = TempPalletId::get().into_account_truncating();
@@ -182,195 +179,12 @@ match_types! {
 	};
 }
 
-pub trait Reserve {
-	/// Returns assets reserve location.
-	fn reserve(&self) -> Option<MultiLocation>;
-}
-
-// Takes the chain part of a MultiAsset
-impl Reserve for MultiAsset {
-	fn reserve(&self) -> Option<MultiLocation> {
-		if let xcmAssetId::Concrete(location) = self.id.clone() {
-			let first_interior = location.first_interior();
-			let parents = location.parent_count();
-			match (parents, first_interior.clone()) {
-				// The only case for non-relay chain will be the chain itself.
-				(0, Some(Parachain(id))) => Some(MultiLocation::new(0, X1(Parachain(id.clone())))),
-				// Only Sibling parachain is recognized.
-				(1, Some(Parachain(id))) => Some(MultiLocation::new(1, X1(Parachain(id.clone())))),
-				// The Relay chain.
-				(1, _) => Some(MultiLocation::parent()),
-				// No other case is allowed for now.
-				_ => None,
-			}
-		} else {
-			None
-		}
-	}
-}
-
-/// A `FilterAssetLocation` implementation. Filters multi native assets whose
-/// reserve is same with `origin`.
-pub struct MultiNativeAsset;
-impl FilterAssetLocation for MultiNativeAsset {
-	fn filter_asset_location(asset: &MultiAsset, origin: &MultiLocation) -> bool {
-		if let Some(ref reserve) = asset.reserve() {
-			if reserve == origin {
-				return true
-			}
-		}
-		false
-	}
-}
-
 pub type Barriers = (
 	TakeWeightCredit,
 	AllowTopLevelPaidExecutionFrom<Everything>,
 	AllowUnpaidExecutionFrom<ParentOrParentsExecutivePlurality>,
 	// ^^^ Parent and its exec plurality get free execution
 );
-
-// We need to know how to charge for incoming assets
-// This takes the first fungible asset, and takes whatever UnitPerSecondGetter establishes
-// UnitsToWeightRatio trait, which needs to be implemented by AssetIdInfoGetter
-pub struct FirstAssetTrader<
-	AssetType: From<MultiLocation> + Clone,
-	AssetIdInfoGetter: UnitsToWeightRatio<AssetType>,
-	R: TakeRevenue,
->(Weight, Option<(MultiLocation, u128, u128)>, PhantomData<(AssetType, AssetIdInfoGetter, R)>);
-impl<
-		AssetType: From<MultiLocation> + Clone,
-		AssetIdInfoGetter: UnitsToWeightRatio<AssetType>,
-		R: TakeRevenue,
-	> WeightTrader for FirstAssetTrader<AssetType, AssetIdInfoGetter, R>
-{
-	fn new() -> Self {
-		FirstAssetTrader(0, None, PhantomData)
-	}
-	fn buy_weight(
-		&mut self,
-		weight: Weight,
-		payment: xcm_executor::Assets,
-	) -> Result<xcm_executor::Assets, XcmError> {
-		let first_asset = payment.fungible_assets_iter().next().ok_or(XcmError::TooExpensive)?;
-
-		// We are only going to check first asset for now. This should be sufficient for simple
-		// token transfers. We will see later if we change this.
-		match (first_asset.id, first_asset.fun) {
-			(xcmAssetId::Concrete(id), Fungibility::Fungible(_)) => {
-				let asset_type: AssetType = id.clone().into();
-				// Shortcut if we know the asset is not supported
-				// This involves the same db read per block, mitigating any attack based on
-				// non-supported assets
-				if !AssetIdInfoGetter::payment_is_supported(asset_type.clone()) {
-					return Err(XcmError::TooExpensive)
-				}
-				if let Some(units_per_second) = AssetIdInfoGetter::get_units_per_second(asset_type)
-				{
-					let amount = units_per_second.saturating_mul(weight as u128) /
-						(WEIGHT_PER_SECOND as u128);
-
-					// We dont need to proceed if the amount is 0
-					// For cases (specially tests) where the asset is very cheap with respect
-					// to the weight needed
-					if amount.is_zero() {
-						return Ok(payment)
-					}
-
-					let required = MultiAsset {
-						fun: Fungibility::Fungible(amount),
-						id: xcmAssetId::Concrete(id.clone()),
-					};
-					let unused =
-						payment.checked_sub(required).map_err(|_| XcmError::TooExpensive)?;
-					self.0 = self.0.saturating_add(weight);
-
-					// In case the asset matches the one the trader already stored before, add
-					// to later refund
-
-					// Else we are always going to substract the weight if we can, but we latter do
-					// not refund it
-
-					// In short, we only refund on the asset the trader first succesfully was able
-					// to pay for an execution
-					let new_asset = match self.1.clone() {
-						Some((prev_id, prev_amount, units_per_second)) =>
-							if prev_id == id {
-								Some((id, prev_amount.saturating_add(amount), units_per_second))
-							} else {
-								None
-							},
-						None => Some((id, amount, units_per_second)),
-					};
-
-					// Due to the trait bound, we can only refund one asset.
-					if let Some(new_asset) = new_asset {
-						self.0 = self.0.saturating_add(weight);
-						self.1 = Some(new_asset);
-					};
-					Ok(unused)
-				} else {
-					Err(XcmError::TooExpensive)
-				}
-			},
-			_ => Err(XcmError::TooExpensive),
-		}
-	}
-
-	fn refund_weight(&mut self, weight: Weight) -> Option<MultiAsset> {
-		if let Some((id, prev_amount, units_per_second)) = self.1.clone() {
-			let weight = weight.min(self.0);
-			self.0 -= weight;
-			let amount = units_per_second * (weight as u128) / (WEIGHT_PER_SECOND as u128);
-			self.1 = Some((id.clone(), prev_amount.saturating_sub(amount), units_per_second));
-			Some(MultiAsset { fun: Fungibility::Fungible(amount), id: xcmAssetId::Concrete(id) })
-		} else {
-			None
-		}
-	}
-}
-
-/// Deal with spent fees, deposit them as dictated by R
-impl<
-		AssetType: From<MultiLocation> + Clone,
-		AssetIdInfoGetter: UnitsToWeightRatio<AssetType>,
-		R: TakeRevenue,
-	> Drop for FirstAssetTrader<AssetType, AssetIdInfoGetter, R>
-{
-	fn drop(&mut self) {
-		if let Some((id, amount, _)) = self.1.clone() {
-			R::take_revenue((id, amount).into());
-		}
-	}
-}
-
-/// XCM fee depositor to which we implement the TakeRevenue trait
-/// It receives a fungibles::Mutate implemented argument, a matcher to convert MultiAsset into
-/// AssetId and amount, and the fee receiver account
-pub struct XcmFeesToAccount<Assets, Matcher, AccountId, ReceiverAccount>(
-	PhantomData<(Assets, Matcher, AccountId, ReceiverAccount)>,
-);
-impl<
-		Assets: Mutate<AccountId>,
-		Matcher: MatchesFungibles<Assets::AssetId, Assets::Balance>,
-		AccountId: Clone,
-		ReceiverAccount: Get<AccountId>,
-	> TakeRevenue for XcmFeesToAccount<Assets, Matcher, AccountId, ReceiverAccount>
-{
-	fn take_revenue(revenue: MultiAsset) {
-		match Matcher::matches_fungibles(&revenue) {
-			Ok((asset_id, amount)) =>
-				if !amount.is_zero() {
-					let ok = Assets::mint_into(asset_id, &ReceiverAccount::get(), amount).is_ok();
-					debug_assert!(ok, "`mint_into` cannot generally fail; qed");
-				},
-			Err(_) => log::debug!(
-				target: "xcm",
-				"take revenue failed matching fungible"
-			),
-		}
-	}
-}
 
 parameter_types! {
 	/// Xcm fees will go to the treasury account
